# backend/docs/swagger.yaml
definitions:
  domain.AuthSessionSummary:
    properties:
      created_at:
        type: string
      current:
        type: boolean
      device_name:
        type: string
      expires_at:
        type: string
      id:
        type: string
      ip_address:
        type: string
      last_seen_at:
        type: string
      user_agent:
        type: string
    type: object
...
  handler.AuthSessionsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.AuthSessionSummary'
        type: array
    type: object
...
  handler.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  handler.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  handler.StatusResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: Customer auth, product catalog, and admin CMS API for the Diagonals
    website backend.
  title: Diagonals API
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a customer and create a new active session for the
        current device or browser
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Login customer account
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      description: Revoke the authenticated session for the current device or browser
        and clear auth cookies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout current session
      tags:
      - Auth
  /api/v1/auth/logout-all:
    post:
      description: Revoke all active sessions for the authenticated user and clear
        auth cookies on the current device
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout all sessions
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      description: Validate the refresh token from HttpOnly cookie against the current
        session, rotate it, and update auth cookies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Refresh access token
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new customer in Go-owned tables and issue access and
        refresh tokens
      parameters:
      - description: Registration payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Register customer account
      tags:
      - Auth
  /api/v1/auth/sessions:
    get:
      description: List all active sessions for the authenticated user across devices
        and browsers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AuthSessionsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List active sessions
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"